# api/openapi.yaml
openapi: 3.1.0
info:
  title: AgentFlow API
  version: 0.1.0
  description: |
    Authoritative HTTP contract shared by generated Go DTOs and the typed
    TypeScript client. Business handlers remain implemented with net/http.
    With AUTH_MODE=oidc, every business operation requires a revocable session
    cookie and owned/granted Workspace entity (401/404 on failure). Mutations also
    require the configured frontend Origin. Health/authentication and preflight
    are exempt. Local mode is unauthenticated trusted development, not an
    alternate authentication mechanism. HTTPS uses __Host-agentflow_session;
    loopback HTTP uses agentflow_session. Archived business writes return 409;
    missing scope resolves to the owner's active default. Workspace management
    is owner-scoped independently of the selected header. Agent/Tool configuration
    writes remain trusted-local operator operations (OIDC users receive 403).
    Resource IDs are resolved inside that authorized Workspace; foreign IDs return
    404. SSE delivery rechecks the session and Workspace on each write, including
    Chat/Continue/Resume. A later access failure emits an error frame using the
    existing ChatStreamEvent error contract and stops delivery, without canceling
    the admitted durable Run. See the resource authorization guide.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/workspaces:
    get:
      operationId: listWorkspaces
      tags: [Workspaces]
      description: Lists only the current user's owned, granted, non-deleted Workspaces, including archived read-only spaces. Local mode uses a persistent server-owned local User. Selected Workspace headers do not scope this endpoint.
      responses:
        '200':
          description: Owned Workspaces.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Workspace' }
    post:
      operationId: createWorkspace
      tags: [Workspaces]
      description: Owner is derived from the session, never accepted from the client. IDs are generated by PostgreSQL; names do not identify resources.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkspaceCreateRequest' }
      responses:
        '201':
          description: Created Workspace.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/workspaces/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, pattern: '^[1-9][0-9]*$' }
    get:
      operationId: getWorkspace
      tags: [Workspaces]
      responses:
        '200':
          description: Owned and granted Workspace.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateWorkspace
      tags: [Workspaces]
      description: Renames, updates description, archives/restores or selects the default. Closing the current default requires an owned active replacement. Unfinished Runs and uncertain effects block archive/delete.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkspaceUpdateRequest' }
      responses:
        '200':
          description: Updated Workspace, preserving its ID.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Last active space, invalid replacement default, unfinished Run or uncertain effect prevents the change.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      operationId: deleteWorkspace
      tags: [Workspaces]
      description: Sets deleted_at. Does not physically erase resource data. Deleted spaces cannot be accessed or implicitly recreated at login.
      parameters:
        - in: query
          name: replacement_workspace_id
          schema: { type: string, pattern: '^[1-9][0-9]*$' }
      responses:
        '204': { description: Soft-deleted. }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Workspace cannot currently be deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    Workspace:
      type: object
      required: [id, owner_user_id, name, description, status, is_default, created_at, updated_at]
      properties:
        id: { type: string, pattern: '^[1-9][0-9]*$', description: Decimal BIGINT identity encoded as a string to preserve precision. }
        owner_user_id: { type: string }
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 2000 }
        status: { type: string, enum: [active, archived] }
        is_default: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time }
    WorkspaceCreateRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 2000 }
    WorkspaceUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 2000 }
        status: { type: string, enum: [active, archived] }
        make_default: { type: boolean }
        replacement_workspace_id: { type: string, pattern: '^[1-9][0-9]*$' }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        source:
          type: string
        category:
          type: string
        retryable:
          type: boolean
        operation:
          type: string
        request_id:
          type: string
...